+++ tool/consumer_contracts/flutter/lib/registry.dart
@@ -1,0 +1,4 @@
+import 'package:factory/factory.dart';
+
+@FactoryRegistry(runtime: FactoryRuntime.flutter)
+void configureFactories() {}