+++ tool/consumer_contracts/flutter/lib/registry.factory.dart
@@ -1,0 +1,9 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+import 'package:factory/factory.dart';
+import 'package:factory_flutter_consumer/dependencies.dart' as factory0;
+
+final appModule = FactoryModule(
+  factories: [factory0.greeting],
+  expose: [factory0.greeting],
+);